+++ schedule/__init__.py
@@ -1,0 +1,3 @@
+"""Scheduler"""
+
+from schedule.scheduler import scheduler
